refactor: Seperate the GUI and Simulator.
