+++ src/pilot/sabre_pilot/factory.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../mcu.hpp"
+#include "../simulator/mcu.hpp"
 #include "sabre/app/app.hpp"
 #include <sabre/factory.hpp>
